test(core-forger): refactor- remove duplicated code
