+++ __tests__/unit/core-forger/__utils__/calculate-active-delegates.ts
@@ -1,0 +1,21 @@
+import { Services } from "@packages/core-kernel";
+import { Wallet } from "@packages/core-state/src/wallets";
+import { Identities } from "@packages/crypto";
+
+export const calculateActiveDelegates = () => {
+    const activeDelegates = [];
+    for (let i = 0; i < 51; i++) {
+        const address = `Delegate-Wallet-${i}`;
+        const wallet = new Wallet(
+            address,
+            new Services.Attributes.AttributeMap(new Services.Attributes.AttributeSet()),
+        );
+
+        wallet.publicKey = Identities.PublicKey.fromPassphrase(address);
+        // @ts-ignore
+        wallet.delegate = { username: `Username: ${address}` };
+        // @ts-ignore
+        activeDelegates.push(wallet);
+    }
+    return activeDelegates;
+};
